feat: make code modular
