+++ sample.env
@@ -1,4 +1,10 @@
+# Bot
 TOKEN=""
+
+# Hooks
 DEBUG=""
+COMM=""
+DIST=""
+
+#Env
 NODE_ENV="development"
-NEW_COMM_HOOK=""
