+++ src/interface/ExtendedClient.ts
@@ -1,0 +1,7 @@
+import { Client, WebhookClient } from "discord.js";
+
+export interface ExtendedClient extends Client {
+  debug: WebhookClient;
+  comm: WebhookClient;
+  dist: WebhookClient;
+}
